Migrated inspection methods to new assistant class
Re-organizing methods for more logical pairing. IngredientTreeUtil will
be used for introspection of the nodes.
