+++ main/java/com/tank/flavorpairer/IngredientTreeAssistant.java
@@ -6,14 +6,14 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import org.assertj.core.util.Lists;
-
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.tank.flavorpairer.object.Ingredient;
 import com.tank.flavorpairer.object.IngredientNode;
 import com.tank.flavorpairer.object.IngredientTree;
 
+/**
+ * Contains assistant methods for {@link IngredientTree}.
+ */
 public class IngredientTreeAssistant {
 	private IngredientTreeAssistant() {
 	}
@@ -152,11 +152,12 @@ public class IngredientTreeAssistant {
 				root.setLeftNode(addIngredientToTree(root.getLeftNode(), ingredientNodeToInsert));
 			}
 
-			if ((getDepth(root.getLeftNode()) - getDepth(root.getRightNode())) > 1) {
+			if ((IngredientTreeUtil.getDepth(root.getLeftNode())
+					- IngredientTreeUtil.getDepth(root.getRightNode())) > 1) {
 				if (root.getLeftNode().getName().compareToIgnoreCase(ingredientNodeToInsert.getName()) > 0) {
-					return rotateWithLeftChild(root);
+					return rotateToRight(root);
 				} else {
-					return doubleWithLeftChild(root);
+					return doubleRotateToRight(root);
 				}
 			}
 		} else if (root.getName().compareToIgnoreCase(ingredientNodeToInsert.getName()) < 0) {
@@ -166,11 +167,12 @@ public class IngredientTreeAssistant {
 				root.setRightNode(addIngredientToTree(root.getRightNode(), ingredientNodeToInsert));
 			}
 
-			if ((getDepth(root.getRightNode()) - getDepth(root.getLeftNode())) > 1) {
+			if ((IngredientTreeUtil.getDepth(root.getRightNode())
+					- IngredientTreeUtil.getDepth(root.getLeftNode())) > 1) {
 				if (root.getRightNode().getName().compareToIgnoreCase(ingredientNodeToInsert.getName()) < 0) {
-					return rotateWithRightChild(root);
+					return rotateToLeft(root);
 				} else {
-					return doubleWithRightChild(root);
+					return doubleRotateToLeft(root);
 				}
 			}
 		} else {
@@ -179,33 +181,53 @@ public class IngredientTreeAssistant {
 		return root;
 	}
 
-	private static IngredientNode rotateWithLeftChild(IngredientNode node) {
-		final IngredientNode leftRootNode = node.getLeftNode();
-		node.setLeftNode(leftRootNode.getRightNode());
-		leftRootNode.setRightNode(node);
-		return leftRootNode;
+	/**
+	 * Rotates the {@link IngredientNode} to the right.
+	 * 
+	 * @param node The {@link IngredientNode} to rotate.
+	 * @return The non-null rotated {@link IngredientNode}.
+	 */
+	private static IngredientNode rotateToRight(IngredientNode node) {
+		final IngredientNode leftNode = node.getLeftNode();
+		node.setLeftNode(leftNode.getRightNode());
+		leftNode.setRightNode(node);
+		return leftNode;
 	}
 
-	private static IngredientNode rotateWithRightChild(IngredientNode node) {
-		final IngredientNode rightRootNode = node.getRightNode();
-		node.setRightNode(rightRootNode.getLeftNode());
-		rightRootNode.setLeftNode(node);
-		return rightRootNode;
+	/**
+	 * Rotates the {@link IngredientNode} to the left.
+	 * 
+	 * @param node The {@link IngredientNode} to rotate.
+	 * @return The non-null rotated {@link IngredientNode}.
+	 */
+	private static IngredientNode rotateToLeft(IngredientNode node) {
+		final IngredientNode rightNode = node.getRightNode();
+		node.setRightNode(rightNode.getLeftNode());
+		rightNode.setLeftNode(node);
+		return rightNode;
 	}
 
-	private static IngredientNode doubleWithLeftChild(IngredientNode node) {
-		node.setLeftNode(rotateWithRightChild(node.getLeftNode()));
-		return rotateWithLeftChild(node);
+	private static IngredientNode doubleRotateToRight(IngredientNode node) {
+		node.setLeftNode(rotateToLeft(node.getLeftNode()));
+		return rotateToRight(node);
 	}
 
-	private static IngredientNode doubleWithRightChild(IngredientNode node) {
-		node.setRightNode(rotateWithLeftChild(node.getRightNode()));
-		return rotateWithRightChild(node);
+	private static IngredientNode doubleRotateToLeft(IngredientNode node) {
+		node.setRightNode(rotateToRight(node.getRightNode()));
+		return rotateToLeft(node);
 	}
 
+	/**
+	 * Deletes the ingredient from the tree.
+	 * 
+	 * @param root       The {@link IngredientNode} to inspect.
+	 * @param ingredient The {@link Ingredient} to delete.
+	 * @return The new {@link IngredientNode} without the given {@link Ingredient}.
+	 *         Will be null if no ingredients remain.
+	 */
 	public static IngredientNode deleteNodeFromTree(IngredientNode root, final Ingredient ingredient) {
 		Preconditions.checkArgument(ingredient != null);
-		final List<Ingredient> ingredients = getIngredients(root);
+		final List<Ingredient> ingredients = IngredientTreeUtil.getIngredients(root);
 		ingredients.remove(ingredient);
 
 		if (ingredients.isEmpty()) {
@@ -215,87 +237,6 @@ public class IngredientTreeAssistant {
 		return constructIngredientTree(ingredients).getRoot();
 	}
 
-	/**
-	 * Determines if the given node is balanced. The depth of the sides cannot be
-	 * greater or equal to 2.
-	 * 
-	 * @param node The root {@link IngredientNode} to evaluate.
-	 * @return True if the tree is balanced, false otherwise.
-	 */
-	private static boolean isTreeBalanced(final IngredientNode node) {
-		return Math.abs(getDepth(node.getLeftNode()) - getDepth(node.getRightNode())) < 2;
-	}
-
-	private static IngredientNode getMostLeftNode(final IngredientNode parentNode, final IngredientNode leftNode) {
-		return leftNode == null ? parentNode : getMostLeftNode(leftNode, leftNode.getLeftNode());
-	}
-
-	private static IngredientNode getMostRightNode(final IngredientNode parentNode, final IngredientNode rightNode) {
-		return rightNode == null ? parentNode : getMostRightNode(rightNode, rightNode.getRightNode());
-	}
-
-	@VisibleForTesting
-	public static IngredientNode findMiddleNode(final IngredientNode rootNode) {
-		final int leftTreeDepth = countNumberOfChildren(rootNode.getLeftNode());
-		final int rightTreeDepth = countNumberOfChildren(rootNode.getRightNode());
-		if (Math.abs(leftTreeDepth - rightTreeDepth) <= 1) {
-			// Since we have an odd number (or same depth), default to root
-			return rootNode;
-		}
-
-		clearAllMarkNodes(rootNode);
-
-		if (leftTreeDepth > rightTreeDepth) {
-			// Since L>R, we will add the root node to R since it will be marked before
-			// starting to mark the other side of L
-			final IngredientNode startingLeftNode = rootNode.getLeftNode();
-			final int roundOfMarks = Math.floorDiv(leftTreeDepth - rightTreeDepth, 2);
-			for (int i = 0; i < roundOfMarks; i++) {
-				final IngredientNode farthestUnmarkedLeft = findFurthestUnmarkedNodeFromRoot(startingLeftNode);
-				farthestUnmarkedLeft.markAsVisited();
-
-				if (rootNode.isVisited()) {
-					// Start marking top of left
-					final IngredientNode closestUnmarkedNode = findClosestUnmarkedNodeFromRoot(startingLeftNode);
-					closestUnmarkedNode.markAsVisited();
-				}
-
-				final IngredientNode farthestUnmarkedRight = findFurthestUnmarkedNodeFromRoot(rootNode.getRightNode());
-				if (farthestUnmarkedRight == null) {
-					rootNode.markAsVisited();
-				} else {
-					farthestUnmarkedRight.markAsVisited();
-				}
-			}
-
-			return findClosestUnmarkedNodeFromRoot(rootNode.getLeftNode());
-		}
-
-		// Since R>L, we will add the root node to L since it will be marked before
-		// starting to mark the other side of R
-		// Mark 2 nodes per round
-		final int roundOfMarks = Math.floorDiv(rightTreeDepth - leftTreeDepth, 2);
-		for (int i = 0; i < roundOfMarks; i++) {
-			final IngredientNode farthestUnmarkedRight = findLargestUnmarkedNode(rootNode);
-			farthestUnmarkedRight.markAsVisited();
-
-			if (rootNode.isVisited()) {
-				// Start marking top of left
-				final IngredientNode closestUnmarkedNode = findClosestUnmarkedNodeFromRoot(rootNode);
-				closestUnmarkedNode.markAsVisited();
-			}
-
-			final IngredientNode farthestUnmarkedLeft = findLargestUnmarkedNode(rootNode.getLeftNode());
-			if (farthestUnmarkedLeft == null) {
-				rootNode.markAsVisited();
-			} else {
-				farthestUnmarkedLeft.markAsVisited();
-			}
-		}
-
-		return findClosestUnmarkedNodeFromRoot(rootNode.getRightNode());
-	}
-
 	private static List<IngredientNode> createIngredients(List<Ingredient> ingredients) {
 		final List<IngredientNode> nodes = ingredients.stream().map(i -> {
 			final IngredientNode node = new IngredientNode(i);
@@ -303,237 +244,5 @@ public class IngredientTreeAssistant {
 			return node;
 		}).collect(Collectors.toList());
 		return nodes;
-	}
-
-	@VisibleForTesting
-	public static IngredientNode findFurthestNodeFromRoot(final IngredientNode ingredientNode) {
-		if ((ingredientNode == null)
-				|| (ingredientNode.getLeftNode() == null && ingredientNode.getRightNode() == null)) {
-			return ingredientNode;
-		}
-
-		final int sizeOfLeft = countNumberOfChildren(ingredientNode.getLeftNode());
-		final int sizeOfRight = countNumberOfChildren(ingredientNode.getRightNode());
-		if (sizeOfLeft == sizeOfRight) {
-			final int depthOfLeft = getDepth(ingredientNode.getLeftNode());
-			final int depthOfRight = getDepth(ingredientNode.getRightNode());
-			if (depthOfLeft == depthOfRight) {
-				return findFurthestNodeFromRoot(ingredientNode.getRightNode());
-			} else if (depthOfLeft > depthOfRight) {
-				return findFurthestNodeFromRoot(ingredientNode.getLeftNode());
-			}
-			return findFurthestNodeFromRoot(ingredientNode.getRightNode());
-		} else if (sizeOfLeft > sizeOfRight) {
-			return findFurthestNodeFromRoot(ingredientNode.getLeftNode());
-		}
-		return findFurthestNodeFromRoot(ingredientNode.getRightNode());
-	}
-
-	/**
-	 * Finds the farthest unvisited node from the given node.
-	 * 
-	 * @param ingredientNode The root {@link IngredientNode} used to determine the
-	 *                       farthest unvisited node.
-	 * @return The unvisited {@link IngredientNode} that is farthest from the root
-	 *         node. Will be null if all nodes, including root, are visited.
-	 */
-	@VisibleForTesting
-	public static IngredientNode findFurthestUnmarkedNodeFromRoot(final IngredientNode ingredientNode) {
-		if ((ingredientNode == null)
-				|| (ingredientNode.getLeftNode() == null && ingredientNode.getRightNode() == null)) {
-			return ingredientNode == null ? null : ingredientNode.isVisited() ? null : ingredientNode;
-		}
-
-		final int sizeOfLeft = countNumberOfUnmarkedChildren(ingredientNode.getLeftNode());
-		final int sizeOfRight = countNumberOfUnmarkedChildren(ingredientNode.getRightNode());
-		if (sizeOfLeft == 0 && sizeOfRight == 0) {
-			return ingredientNode.isVisited() ? null : ingredientNode;
-		}
-
-		if (sizeOfLeft == sizeOfRight) {
-			final int depthOfLeft = getUnmarkedDepth(ingredientNode.getLeftNode());
-			final int depthOfRight = getUnmarkedDepth(ingredientNode.getRightNode());
-			if (depthOfLeft == depthOfRight) {
-				return findFurthestUnmarkedNodeFromRoot(ingredientNode.getRightNode());
-			} else if (depthOfLeft > depthOfRight) {
-				return findFurthestUnmarkedNodeFromRoot(ingredientNode.getLeftNode());
-			}
-			return findFurthestUnmarkedNodeFromRoot(ingredientNode.getRightNode());
-		} else if (sizeOfLeft > sizeOfRight) {
-			return findFurthestUnmarkedNodeFromRoot(ingredientNode.getLeftNode());
-		}
-		return findFurthestUnmarkedNodeFromRoot(ingredientNode.getRightNode());
-	}
-
-	@VisibleForTesting
-	public static IngredientNode findLargestUnmarkedNode(final IngredientNode ingredientNode) {
-		if ((ingredientNode == null)
-				|| (ingredientNode.getLeftNode() == null && ingredientNode.getRightNode() == null)) {
-			return ingredientNode == null ? null : ingredientNode.isVisited() ? null : ingredientNode;
-		}
-
-		final int sizeOfLeft = countNumberOfUnmarkedChildren(ingredientNode.getLeftNode());
-		final int sizeOfRight = countNumberOfUnmarkedChildren(ingredientNode.getRightNode());
-		if (sizeOfLeft == 0 && sizeOfRight == 0) {
-			return ingredientNode.isVisited() ? null : ingredientNode;
-		}
-
-		if (sizeOfRight > 0) {
-			return findLargestUnmarkedNode(ingredientNode.getRightNode());
-		} else if (!ingredientNode.isVisited()) {
-			return ingredientNode;
-		} else {
-			return findLargestUnmarkedNode(ingredientNode.getLeftNode());
-		}
-	}
-
-	@VisibleForTesting
-	public static IngredientNode findClosestUnmarkedNodeFromRoot(final IngredientNode ingredientNode) {
-		if ((ingredientNode == null)
-				|| (ingredientNode.getLeftNode() == null && ingredientNode.getRightNode() == null)) {
-			return ingredientNode == null ? null : ingredientNode.isVisited() ? null : ingredientNode;
-		}
-
-		if (!ingredientNode.isVisited()) {
-			return ingredientNode;
-		}
-
-		final int sizeOfLeft = countNumberOfUnmarkedChildren(ingredientNode.getLeftNode());
-		final int sizeOfRight = countNumberOfUnmarkedChildren(ingredientNode.getRightNode());
-		if (sizeOfLeft == 0 && sizeOfRight == 0) {
-			return ingredientNode.isVisited() ? null : ingredientNode;
-		}
-
-		final int depthOfLeft = getUnmarkedDepth(ingredientNode.getLeftNode());
-		final int depthOfRight = getUnmarkedDepth(ingredientNode.getRightNode());
-
-		if (sizeOfLeft == sizeOfRight) {
-			if (depthOfLeft == depthOfRight) {
-				return findClosestUnmarkedNodeFromRoot(ingredientNode.getRightNode());
-			} else if (depthOfLeft > depthOfRight) {
-				return findClosestUnmarkedNodeFromRoot(ingredientNode.getLeftNode());
-			}
-			return findClosestUnmarkedNodeFromRoot(ingredientNode.getRightNode());
-		} else if (sizeOfLeft > sizeOfRight) {
-			return findClosestUnmarkedNodeFromRoot(ingredientNode.getLeftNode());
-		}
-		return findClosestUnmarkedNodeFromRoot(ingredientNode.getRightNode());
-	}
-
-	@VisibleForTesting
-	public static int countNumberOfChildren(IngredientNode ingredientNode) {
-		if (ingredientNode == null) {
-			return 0;
-		}
-
-		if (ingredientNode.getLeftNode() == null && ingredientNode.getRightNode() == null) {
-			return 1;
-		}
-
-		final int sizeOfLeft = countNumberOfChildren(ingredientNode.getLeftNode());
-		final int sizeOfRight = countNumberOfChildren(ingredientNode.getRightNode());
-		return 1 + sizeOfLeft + sizeOfRight;
-	}
-
-	@VisibleForTesting
-	public static int countNumberOfUnmarkedChildren(IngredientNode ingredientNode) {
-		if (ingredientNode == null) {
-			return 0;
-		}
-
-		if (ingredientNode.isVisited() && ingredientNode.getLeftNode() == null
-				&& ingredientNode.getRightNode() == null) {
-			return 0;
-		}
-
-		if (!ingredientNode.isVisited() && ingredientNode.getLeftNode() == null
-				&& ingredientNode.getRightNode() == null) {
-			return 1;
-		}
-
-		final int sizeOfLeft = countNumberOfUnmarkedChildren(ingredientNode.getLeftNode());
-		final int sizeOfRight = countNumberOfUnmarkedChildren(ingredientNode.getRightNode());
-		return (ingredientNode.isVisited() ? 0 : 1) + sizeOfLeft + sizeOfRight;
-	}
-
-	@VisibleForTesting
-	public static int getDepth(IngredientNode ingredientNode) {
-		if (ingredientNode == null) {
-			return 0;
-		}
-
-		if (ingredientNode.getLeftNode() == null && ingredientNode.getRightNode() == null) {
-			return 1;
-		}
-
-		final int sizeOfLeft = getDepth(ingredientNode.getLeftNode(), 1);
-		final int sizeOfRight = getDepth(ingredientNode.getRightNode(), 1);
-		return Integer.max(sizeOfLeft, sizeOfRight);
-	}
-
-	private static int getDepth(IngredientNode ingredientNode, int size) {
-		if (ingredientNode == null) {
-			return size;
-		}
-
-		++size;
-		final int sizeOfLeft = getDepth(ingredientNode.getLeftNode(), size);
-		final int sizeOfRight = getDepth(ingredientNode.getRightNode(), size);
-		return Integer.max(sizeOfLeft, sizeOfRight);
-	}
-
-	@VisibleForTesting
-	public static int getUnmarkedDepth(IngredientNode ingredientNode) {
-		if (ingredientNode == null || ingredientNode.isVisited()) {
-			return 0;
-		}
-
-		// May not need anymore?
-//		final boolean isLeftDead = ingredientNode.getLeftNode() == null || ingredientNode.getLeftNode().isVisited();
-//		final boolean isRightDead = ingredientNode.getRightNode() == null || ingredientNode.getRightNode().isVisited();
-//		if (ingredientNode.isVisited() && isLeftDead && isRightDead) {
-//			return 0;
-//		}
-
-		final int initialStartingDepth = ingredientNode.isVisited() ? 0 : 1;
-		final int sizeOfLeft = getUnmarkedDepth(ingredientNode.getLeftNode(), initialStartingDepth);
-		final int sizeOfRight = getUnmarkedDepth(ingredientNode.getRightNode(), initialStartingDepth);
-		return Integer.max(sizeOfLeft, sizeOfRight);
-	}
-
-	private static int getUnmarkedDepth(IngredientNode ingredientNode, int size) {
-		if (ingredientNode == null) {
-			return size;
-		}
-
-		if (!ingredientNode.isVisited()) {
-			++size;
-		}
-
-		final int sizeOfLeft = getUnmarkedDepth(ingredientNode.getLeftNode(), size);
-		final int sizeOfRight = getUnmarkedDepth(ingredientNode.getRightNode(), size);
-		return Integer.max(sizeOfLeft, sizeOfRight);
-	}
-
-	public static List<Ingredient> getIngredients(IngredientNode ingredientNode) {
-		if (ingredientNode == null) {
-			return Lists.emptyList();
-		}
-
-		final List<Ingredient> ingredients = Lists.newArrayList();
-		ingredients.add(ingredientNode.getIngredient());
-		ingredients.addAll(getIngredients(ingredientNode.getLeftNode()));
-		ingredients.addAll(getIngredients(ingredientNode.getRightNode()));
-		return ingredients;
-	}
-
-	private static void clearAllMarkNodes(IngredientNode ingredientNode) {
-		if (ingredientNode == null) {
-			return;
-		}
-
-		ingredientNode.unmarkAsVisited();
-		clearAllMarkNodes(ingredientNode.getLeftNode());
-		clearAllMarkNodes(ingredientNode.getRightNode());
 	}
 }
