+++ test/java/com/tank/flavorpairer/IngredientTreeUtil_CountNumberOfChildren_Test.java
@@ -8,13 +8,13 @@ import org.junit.Test;
 import com.tank.flavorpairer.object.Ingredient;
 import com.tank.flavorpairer.object.IngredientNode;
 
-public class IngredientTreeAssistant_CountNumberOfChildren_Test {
+public class IngredientTreeUtil_CountNumberOfChildren_Test {
 	@Test
 	public void testOnlyRoot() {
 		// Tree
 		// A
 		final IngredientNode rootNode = createIngredientNode(Ingredient.APPLE);
-		Assert.assertEquals(1, IngredientTreeAssistant.countNumberOfChildren(rootNode));
+		Assert.assertEquals(1, IngredientTreeUtil.countNumberOfChildren(rootNode));
 	}
 
 	@Test
@@ -24,7 +24,7 @@ public class IngredientTreeAssistant_CountNumberOfChildren_Test {
 		// A
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE),
 				(IngredientNode) null);
-		Assert.assertEquals(2, IngredientTreeAssistant.countNumberOfChildren(rootNode));
+		Assert.assertEquals(2, IngredientTreeUtil.countNumberOfChildren(rootNode));
 	}
 
 	@Test
@@ -34,7 +34,7 @@ public class IngredientTreeAssistant_CountNumberOfChildren_Test {
 		// ..B
 		final IngredientNode rootNode = createIngredientNode(Ingredient.APPLE, (IngredientNode) null,
 				createIngredientNode(Ingredient.BASIL));
-		Assert.assertEquals(2, IngredientTreeAssistant.countNumberOfChildren(rootNode));
+		Assert.assertEquals(2, IngredientTreeUtil.countNumberOfChildren(rootNode));
 	}
 
 	@Test
@@ -44,7 +44,7 @@ public class IngredientTreeAssistant_CountNumberOfChildren_Test {
 		// A C
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE),
 				createIngredientNode(Ingredient.CINNAMON));
-		Assert.assertEquals(3, IngredientTreeAssistant.countNumberOfChildren(rootNode));
+		Assert.assertEquals(3, IngredientTreeUtil.countNumberOfChildren(rootNode));
 	}
 
 	@Test
@@ -56,7 +56,7 @@ public class IngredientTreeAssistant_CountNumberOfChildren_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE), (IngredientNode) null),
 				createIngredientNode(Ingredient.ZUCCHINI));
-		Assert.assertEquals(4, IngredientTreeAssistant.countNumberOfChildren(rootNode));
+		Assert.assertEquals(4, IngredientTreeUtil.countNumberOfChildren(rootNode));
 	}
 
 	@Test
@@ -68,7 +68,7 @@ public class IngredientTreeAssistant_CountNumberOfChildren_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL), createIngredientNode(Ingredient.FENNEL, (IngredientNode) null,
 						createIngredientNode(Ingredient.ZUCCHINI)));
-		Assert.assertEquals(4, IngredientTreeAssistant.countNumberOfChildren(rootNode));
+		Assert.assertEquals(4, IngredientTreeUtil.countNumberOfChildren(rootNode));
 	}
 
 	@Test
@@ -80,6 +80,6 @@ public class IngredientTreeAssistant_CountNumberOfChildren_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL), createIngredientNode(Ingredient.FENNEL,
 						createIngredientNode(Ingredient.CORIANDER), createIngredientNode(Ingredient.ZUCCHINI)));
-		Assert.assertEquals(5, IngredientTreeAssistant.countNumberOfChildren(rootNode));
+		Assert.assertEquals(5, IngredientTreeUtil.countNumberOfChildren(rootNode));
 	}
 }
