+++ test/java/com/tank/flavorpairer/IngredientTreeUtil_FindFurthestNodeFromRoot_Test.java
@@ -8,10 +8,10 @@ import org.junit.Test;
 import com.tank.flavorpairer.object.Ingredient;
 import com.tank.flavorpairer.object.IngredientNode;
 
-public class IngredientTreeAssistant_FindFurthestNodeFromRoot_Test {
+public class IngredientTreeUtil_FindFurthestNodeFromRoot_Test {
 	@Test
 	public void testEmpty() {
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findFurthestNodeFromRoot(null);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findFurthestNodeFromRoot(null);
 		Assert.assertEquals(null, ingredientNode);
 	}
 
@@ -20,7 +20,7 @@ public class IngredientTreeAssistant_FindFurthestNodeFromRoot_Test {
 		// Tree
 		// A
 		final IngredientNode rootNode = createIngredientNode(Ingredient.APPLE);
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findFurthestNodeFromRoot(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findFurthestNodeFromRoot(rootNode);
 		Assert.assertEquals(Ingredient.APPLE, ingredientNode.getIngredient());
 	}
 
@@ -31,7 +31,7 @@ public class IngredientTreeAssistant_FindFurthestNodeFromRoot_Test {
 		// A
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE),
 				(IngredientNode) null);
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findFurthestNodeFromRoot(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findFurthestNodeFromRoot(rootNode);
 		Assert.assertEquals(Ingredient.APPLE, ingredientNode.getIngredient());
 	}
 
@@ -42,7 +42,7 @@ public class IngredientTreeAssistant_FindFurthestNodeFromRoot_Test {
 		// ..C
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, (IngredientNode) null,
 				createIngredientNode(Ingredient.CINNAMON));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findFurthestNodeFromRoot(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findFurthestNodeFromRoot(rootNode);
 		Assert.assertEquals(Ingredient.CINNAMON, ingredientNode.getIngredient());
 	}
 
@@ -53,7 +53,7 @@ public class IngredientTreeAssistant_FindFurthestNodeFromRoot_Test {
 		// A C
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE),
 				createIngredientNode(Ingredient.CINNAMON));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findFurthestNodeFromRoot(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findFurthestNodeFromRoot(rootNode);
 		Assert.assertEquals(Ingredient.CINNAMON, ingredientNode.getIngredient());
 	}
 
@@ -66,7 +66,7 @@ public class IngredientTreeAssistant_FindFurthestNodeFromRoot_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE), (IngredientNode) null),
 				createIngredientNode(Ingredient.ZUCCHINI));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findFurthestNodeFromRoot(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findFurthestNodeFromRoot(rootNode);
 		Assert.assertEquals(Ingredient.APPLE, ingredientNode.getIngredient());
 	}
 
@@ -79,7 +79,7 @@ public class IngredientTreeAssistant_FindFurthestNodeFromRoot_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL), createIngredientNode(Ingredient.FENNEL, (IngredientNode) null,
 						createIngredientNode(Ingredient.ZUCCHINI)));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findFurthestNodeFromRoot(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findFurthestNodeFromRoot(rootNode);
 		Assert.assertEquals(Ingredient.ZUCCHINI, ingredientNode.getIngredient());
 	}
 
@@ -92,7 +92,7 @@ public class IngredientTreeAssistant_FindFurthestNodeFromRoot_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL), createIngredientNode(Ingredient.FENNEL,
 						createIngredientNode(Ingredient.CORIANDER), createIngredientNode(Ingredient.ZUCCHINI)));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findFurthestNodeFromRoot(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findFurthestNodeFromRoot(rootNode);
 		Assert.assertEquals(Ingredient.ZUCCHINI, ingredientNode.getIngredient());
 	}
 }
