+++ test/java/com/tank/flavorpairer/IngredientTreeUtil_FindMiddleNode_Test.java
@@ -8,13 +8,13 @@ import org.junit.Test;
 import com.tank.flavorpairer.object.Ingredient;
 import com.tank.flavorpairer.object.IngredientNode;
 
-public class IngredientTreeAssistant_FindMiddleNode_Test {
+public class IngredientTreeUtil_FindMiddleNode_Test {
 	@Test
 	public void testOneElement_Root() {
 		// Tree
 		// A
 		final IngredientNode rootNode = createIngredientNode(Ingredient.APPLE);
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.APPLE, ingredientNode.getIngredient());
 	}
 
@@ -25,7 +25,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 		// A
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE),
 				(IngredientNode) null);
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.BASIL, ingredientNode.getIngredient());
 	}
 
@@ -36,7 +36,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 		// A
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE),
 				(IngredientNode) null);
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.BASIL, ingredientNode.getIngredient());
 	}
 
@@ -47,7 +47,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 		// A C
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE),
 				createIngredientNode(Ingredient.CINNAMON));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.BASIL, ingredientNode.getIngredient());
 	}
 
@@ -60,7 +60,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL, createIngredientNode(Ingredient.APPLE), (IngredientNode) null),
 				createIngredientNode(Ingredient.ZUCCHINI));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.CINNAMON, ingredientNode.getIngredient());
 	}
 
@@ -73,7 +73,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL), createIngredientNode(Ingredient.FENNEL, (IngredientNode) null,
 						createIngredientNode(Ingredient.ZUCCHINI)));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.CINNAMON, ingredientNode.getIngredient());
 	}
 
@@ -86,7 +86,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON, (IngredientNode) null,
 				createIngredientNode(Ingredient.FENNEL, (IngredientNode) null,
 						createIngredientNode(Ingredient.ZUCCHINI)));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.FENNEL, ingredientNode.getIngredient());
 	}
 
@@ -99,7 +99,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 		final IngredientNode rootNode = createIngredientNode(Ingredient.BASIL, (IngredientNode) null,
 				createIngredientNode(Ingredient.FENNEL, createIngredientNode(Ingredient.CINNAMON),
 						(IngredientNode) null));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.CINNAMON, ingredientNode.getIngredient());
 	}
 
@@ -114,7 +114,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 				createIngredientNode(Ingredient.CINNAMON));
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CORIANDER, leftNode,
 				createIngredientNode(Ingredient.ZUCCHINI));
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.BASIL, ingredientNode.getIngredient());
 	}
 
@@ -129,7 +129,7 @@ public class IngredientTreeAssistant_FindMiddleNode_Test {
 				createIngredientNode(Ingredient.CORIANDER), createIngredientNode(Ingredient.ZUCCHINI));
 		final IngredientNode rootNode = createIngredientNode(Ingredient.CINNAMON,
 				createIngredientNode(Ingredient.BASIL), rightNode);
-		final IngredientNode ingredientNode = IngredientTreeAssistant.findMiddleNode(rootNode);
+		final IngredientNode ingredientNode = IngredientTreeUtil.findMiddleNode(rootNode);
 		Assert.assertEquals(Ingredient.FENNEL, ingredientNode.getIngredient());
 	}
 }
